Moved the VQ and VQ Trainer classes to model.py
